+++ tembo-operator/src/extensions/mod.rs
@@ -1,3 +1,4 @@
+pub mod database_queries;
 pub mod types;
 
 use crate::{
@@ -11,10 +12,11 @@ use regex::Regex;
 
 
 use serde_json::json;
-use std::{collections::HashMap, sync::Arc};
+use std::sync::Arc;
 
 use tokio::time::Duration;
 use tracing::{debug, error, info, warn};
+
 use types::{
     Extension, ExtensionInstallLocation, ExtensionInstallLocationStatus, ExtensionStatus, InstallStatus,
     TrunkInstall, TrunkInstallStatus,
@@ -24,79 +26,6 @@ use types::{
 lazy_static! {
     static ref VALID_INPUT: Regex = Regex::new(r"^[a-zA-Z]([a-zA-Z0-9]*[-_]?)*[a-zA-Z0-9]+$").unwrap();
 }
-
-#[derive(Debug)]
-pub struct ExtRow {
-    pub name: String,
-    pub description: String,
-    pub version: String,
-    pub enabled: bool,
-    pub schema: String,
-}
-
-const LIST_DATABASES_QUERY: &str = r#"SELECT datname FROM pg_database WHERE datistemplate = false;"#;
-const LIST_EXTENSIONS_QUERY: &str = r#"select
-distinct on
-(name) *
-from
-(
-select
-    name,
-    version,
-    enabled,
-    schema,
-    description
-from
-    (
-    select
-        t0.extname as name,
-        t0.extversion as version,
-        true as enabled,
-        t1.nspname as schema,
-        comment as description
-    from
-        (
-        select
-            extnamespace,
-            extname,
-            extversion
-        from
-            pg_extension
-) t0,
-        (
-        select
-            oid,
-            nspname
-        from
-            pg_namespace
-) t1,
-        (
-        select
-            name,
-            comment
-        from
-            pg_catalog.pg_available_extensions
-) t2
-    where
-        t1.oid = t0.extnamespace
-        and t2.name = t0.extname 
-) installed
-union
-select
-    name,
-    default_version as version,
-    false as enabled,
-    'public' as schema,
-    comment as description
-from
-    pg_catalog.pg_available_extensions
-order by
-    enabled asc 
-) combined
-order by
-name asc,
-enabled desc
-"#;
 
 pub async fn reconcile_trunk_installs(
     cdb: &CoreDB,
@@ -378,66 +307,6 @@ pub fn check_input(input: &str) -> bool {
     VALID_INPUT.is_match(input)
 }
 
-/// returns all the databases in an instance
-pub async fn list_databases(cdb: &CoreDB, ctx: Arc<Context>) -> Result<Vec<String>, Action> {
-    let _client = ctx.client.clone();
-    let psql_out = cdb
-        .psql(LIST_DATABASES_QUERY.to_owned(), "postgres".to_owned(), ctx)
-        .await?;
-    let result_string = psql_out.stdout.unwrap();
-    Ok(parse_databases(&result_string))
-}
-
-fn parse_databases(psql_str: &str) -> Vec<String> {
-    let mut databases = vec![];
-    for line in psql_str.lines().skip(2) {
-        let fields: Vec<&str> = line.split('|').map(|s| s.trim()).collect();
-        if fields.is_empty()
-            || fields[0].is_empty()
-            || fields[0].contains("rows)")
-            || fields[0].contains("row)")
-        {
-            debug!("Done:{:?}", fields);
-            continue;
-        }
-        databases.push(fields[0].to_string());
-    }
-    let num_databases = databases.len();
-    info!("Found {} databases", num_databases);
-    databases
-}
-
-/// lists all extensions in a single database
-pub async fn list_extensions(cdb: &CoreDB, ctx: Arc<Context>, database: &str) -> Result<Vec<ExtRow>, Action> {
-    let psql_out = cdb
-        .psql(LIST_EXTENSIONS_QUERY.to_owned(), database.to_owned(), ctx)
-        .await?;
-    let result_string = psql_out.stdout.unwrap();
-    Ok(parse_extensions(&result_string))
-}
-
-fn parse_extensions(psql_str: &str) -> Vec<ExtRow> {
-    let mut extensions = vec![];
-    for line in psql_str.lines().skip(2) {
-        let fields: Vec<&str> = line.split('|').map(|s| s.trim()).collect();
-        if fields.len() < 5 {
-            debug!("Done:{:?}", fields);
-            continue;
-        }
-        let package = ExtRow {
-            name: fields[0].to_owned(),
-            version: fields[1].to_owned(),
-            enabled: fields[2] == "t",
-            schema: fields[3].to_owned(),
-            description: fields[4].to_owned(),
-        };
-        extensions.push(package);
-    }
-    let num_extensions = extensions.len();
-    debug!("Found {} extensions", num_extensions);
-    extensions
-}
-
 fn get_location_status(
     cdb: &CoreDB,
     extension_name: &str,
@@ -679,7 +548,7 @@ async fn update_extension_location_in_coredb_status(
 }
 
 async fn create_or_drop_extensions(cdb: &CoreDB, ctx: Arc<Context>) -> Result<Vec<ExtensionStatus>, Action> {
-    let all_actually_installed_extensions = get_all_extensions(cdb, ctx.clone()).await?;
+    let all_actually_installed_extensions = database_queries::get_all_extensions(cdb, ctx.clone()).await?;
     let mut ext_status_updates = determine_updated_extensions_status(cdb, all_actually_installed_extensions);
     let patch_status = json!({
         "apiVersion": "coredb.io/v1alpha1",
@@ -754,43 +623,6 @@ async fn create_or_drop_extensions(cdb: &CoreDB, ctx: Arc<Context>) -> Result<Ve
     Ok(ext_status_updates)
 }
 
-/// list databases then get all extensions from each database
-pub async fn get_all_extensions(cdb: &CoreDB, ctx: Arc<Context>) -> Result<Vec<Extension>, Action> {
-    let databases = list_databases(cdb, ctx.clone()).await?;
-    debug!("databases: {:?}", databases);
-
-    let mut ext_hashmap: HashMap<(String, String), Vec<ExtensionInstallLocation>> = HashMap::new();
-    // query every database for extensions
-    // transform results by extension name, rather than by database
-    for db in databases {
-        let extensions = list_extensions(cdb, ctx.clone(), &db).await?;
-        for ext in extensions {
-            let extlocation = ExtensionInstallLocation {
-                database: db.clone(),
-                version: Some(ext.version),
-                enabled: ext.enabled,
-                schema: ext.schema,
-            };
-            ext_hashmap
-                .entry((ext.name, ext.description))
-                .or_insert_with(Vec::new)
-                .push(extlocation);
-        }
-    }
-
-    let mut ext_spec: Vec<Extension> = Vec::new();
-    for ((extname, extdescr), ext_locations) in &ext_hashmap {
-        ext_spec.push(Extension {
-            name: extname.clone(),
-            description: Some(extdescr.clone()),
-            locations: ext_locations.clone(),
-        });
-    }
-    // put them in order
-    ext_spec.sort_by_key(|e| e.name.clone());
-    Ok(ext_spec)
-}
-
 /// reconcile extensions between the spec and the database
 pub async fn reconcile_extensions(
     coredb: &CoreDB,
@@ -806,6 +638,7 @@ pub async fn reconcile_extensions(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::extensions::database_queries::{parse_databases, parse_extensions};
 
     #[test]
     fn test_parse_databases() {
